+++ saleor/graphql/app/types.py
@@ -108,7 +108,6 @@ class App(CountableDjangoObjectType):
         return format_permissions_for_display(permissions)
 
     @staticmethod
-    @traced_resolver
     def resolve_tokens(root: models.App, _info, **_kwargs):
         return root.tokens.all()  # type: ignore
 
@@ -122,6 +121,7 @@ class App(CountableDjangoObjectType):
         return root.webhooks.all()
 
     @staticmethod
+    @traced_resolver
     def resolve_access_token(root: models.App, info):
         return resolve_access_token(info, root)