+++ saleor/graphql/order/tests/test_fulfillment.py
@@ -84,6 +84,7 @@ def test_order_fulfill(
     }
     mock_create_fulfillments.assert_called_once_with(
         staff_user,
+        None,
         order,
         fulfillment_lines_for_warehouses,
         ANY,
@@ -139,6 +140,7 @@ def test_order_fulfill_as_app(
     }
     mock_create_fulfillments.assert_called_once_with(
         AnonymousUser(),
+        app_api_client.app,
         order,
         fulfillment_lines_for_warehouses,
         ANY,
@@ -202,7 +204,13 @@ def test_order_fulfill_many_warehouses(
     }
 
     mock_create_fulfillments.assert_called_once_with(
-        staff_user, order, fulfillment_lines_for_warehouses, ANY, True, accepted=True
+        staff_user,
+        None,
+        order,
+        fulfillment_lines_for_warehouses,
+        ANY,
+        True,
+        accepted=True,
     )
 
 
@@ -244,7 +252,13 @@ def test_order_fulfill_without_notification(
         str(warehouse.pk): [{"order_line": order_line, "quantity": 1}]
     }
     mock_create_fulfillments.assert_called_once_with(
-        staff_user, order, fulfillment_lines_for_warehouses, ANY, False, accepted=True
+        staff_user,
+        None,
+        order,
+        fulfillment_lines_for_warehouses,
+        ANY,
+        False,
+        accepted=True,
     )
 
 
@@ -302,7 +316,13 @@ def test_order_fulfill_lines_with_empty_quantity(
         str(warehouse.pk): [{"order_line": order_line2, "quantity": 2}]
     }
     mock_create_fulfillments.assert_called_once_with(
-        staff_user, order, fulfillment_lines_for_warehouses, ANY, True, accepted=True
+        staff_user,
+        None,
+        order,
+        fulfillment_lines_for_warehouses,
+        ANY,
+        True,
+        accepted=True,
     )